requirements.txt
streamlit>=1.32,<2.0
langchain
faiss-cpu
transformers
sentence-transformers
openai
pandas
